Slight reorganization to add SolCalcGUI.py to path during installation.
