+++ setup.py
@@ -1,4 +1,22 @@
+import sys
+import os
 import setuptools
+from setuptools.command.develop import develop
+from setuptools.command.install import install
+sys.path.append(os.path.dirname(__file__)+'/scripts/')
+from cfgs import run_cfgs
+
+class PostDevelopCommand(develop):
+    """Post-installation for development mode."""
+    def run(self):
+        develop.run(self)
+        run_cfgs()
+
+class PostInstallCommand(install):
+    """Post-installation for installation mode."""
+    def run(self):
+        install.run(self)
+        run_cfgs()
 
 with open("README.md", "r") as fh:
     long_description = fh.read()
@@ -19,5 +37,11 @@ setuptools.setup(
     python_requires='>=3.6',
     include_package_data=True,
     zip_safe=False,
-    scripts=['scripts/cfgs.py']
+    # cmdclass={
+    #     'develop': PostDevelopCommand,
+    #     'install': PostInstallCommand,
+    # },
+    scripts=['scripts/SolCalc_GUI/SolCalcGUI.py']
 )
+
+
